foodhoover_scrape/requirements.txt
Flask==1.1.2
asyncio==3.4.3
aiohttp==3.7.4.post0
pandas==1.3.1
SQLAlchemy==1.4.11
psycopg2==2.8.6
google-cloud-bigquery==2.7.0
google-cloud-bigquery-storage[pandas,pyarrow]==2.4.0
google-api-python-client==2.3.0
google-cloud-storage==1.38.0
oauth2client==4.1.3